esonic
ESONIC INDIA is an Indian company for more than a decade of experience in manufacturing different kinds of mobile accessories. Now entered the tempered glass market in india. started its online and offline campaign with young and experienced entrepreneurs. we are in wholesale and retail market both. Our mission is to manufacture strong, non breakable tempered glasses that are completely make in inda. and deliver those 50-100rs tempered glasses at affordable prices to you and your business.
esonicindia.com